refactor to move generic utilities out of habitat/datasets/rearrange
